pyproject.toml
[project]
name = "auto-godot"
version = "1.2.0"
description = "Agent-native CLI for Godot Engine"
license = "Apache-2.0"
requires-python = ">=3.12"
dependencies = [
    "click>=8.3",
    "rich-click>=1.9",
]

[project.optional-dependencies]
image = ["Pillow>=12.0"]
dev = [
    "pytest>=9.0",
    "pytest-asyncio>=0.24",
    "pytest-cov>=7.0",
    "ruff",
    "mypy>=1.19",
]

[project.scripts]
auto-godot = "auto_godot.cli:cli"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/auto_godot"]

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"
markers = [
    "requires_godot: marks tests that need Godot binary on PATH",
]

[tool.ruff]
target-version = "py312"
line-length = 99

[tool.ruff.lint]
select = ["E", "F", "W", "I", "UP", "B", "SIM", "TCH"]
ignore = ["E501"]

[tool.mypy]
python_version = "3.12"
strict = true
warn_return_any = true
warn_unused_configs = true
